Update configureAux.sh
When using the device ID, the selection is not persistent after removing the vita from the dock and reconnecting it. (i.e. audio works on initial setup but reconnecting the vita produces no sound.) This is because the the device ID of the usb audio input adapter changes between reconnections. Changing this to use the device name instead fixed this issue in my setup. I was using the audio out from the official Vita 1000 dock from Sony, and the Sonos stereo line-in aux to usb-c adapter, with a usb-c to usb-a adapter on my rpi 4 when the issue occured.
(This is my first github commit, sorry if I did any of it wrong)